server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyserial==3.5
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
aiohttp==3.9.1
